Move airtable into its own package.
