+++ loc/schema.py
@@ -5,7 +5,7 @@ from django.forms import ModelForm
 
 from project.util.session_mutation import SessionFormMutation
 from . import forms, models
-from project import airtable
+from airtable.sync import sync_user as sync_user_with_airtable
 
 
 class OneToOneUserModelFormMutation(SessionFormMutation):
@@ -98,7 +98,7 @@ class LetterRequest(OneToOneUserModelFormMutation):
         request = info.context
         lr = form.save()
         if lr.mail_choice == 'WE_WILL_MAIL':
-            airtable.sync_user(request.user)
+            sync_user_with_airtable(request.user)
             lr.user.send_sms(
                 f"We'll follow up with you about your letter of complaint "
                 f"in about a week, {lr.user.first_name}.",
